linked all templates together

+++ src/main/java/worldMap/ContinentRepository.java
@@ -5,7 +5,7 @@ package worldMap;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.ui.Model;
 
-public interface ContinentRepository extends CrudRepository<Continent, Integer> {
+public interface ContinentRepository extends CrudRepository<Continent, Long> {
